CandyHapsinator

hapsinatoR is used to randomly select candies from a bland-selv-slik rack

hapsinatoR(
  pricePer100g = NULL,
  snollerMoney = NULL,
  numRows = NULL,
  numCols = NULL,
  handfulSize = c("medium", "small", "large")
)

Arguments

pricePer100g

number denoting the price of 100g of candy

snollerMoney

number denoting the money to spend on candy

numRows

integer denoting the number of rows in the candy shelf

numCols

integer denoting the number of columns in the candy shelf

handfulSize

factor denoting handful size (small, medium or large)

Value

hapsinatoR returns an object of class "hapsinatoR", which is a list containing the following components:

candyShelf

(x,y) coordinates with positions to grab candy

numHaps

Number of candy hapsers to make

gramsToMix

Number denoting the amount of candy (in grams) you aim to mix

uniqueCandies

Number of unique pieces of candy

hapsSizes

Vector of length numHaps denoting the sizes of each haps

Details

handfulSize has three possible arguments:

"small"

handfulSize is normally distributed with mean 17 and sd = 3

"medium"

handfulSize is normally distributed with mean 24 and sd = 3

"large"

handfulSize is normally distributed with mean 31 and sd = 3

Examples

## Mix candy for 94kr with a price of 12kr/100g, 5 rows and 12 columns of candy, small handful size.
hapsinatoR(12, 94, 5, 12, "small")
#> A hapsinatoR object
#> Number of candies: 60 
#> Tile statuses:
#> 
#>  GRAB don't 
#>    30    30 
#> Coordinates of candies to GRAB:
#> (3,1)  (2,2)  (3,2)  (4,2)  (3,3)  (4,3)  (2,4)  (4,4)  (4,5)  (1,6)  (3,6)  (4,7)  (2,8)  (3,8)  (4,8)  (2,9)  (3,9)  (4,9)  (5,9)  (2,10)  (3,10)  (4,10)  (1,11)  (2,11)  (3,11)  (4,11)  (5,11)  (2,12)  (4,12)  (5,12)